Modernize `array` package via `{-# LANGUAGE #-}` pragmas
This converts the Cabal field `default-extensions`into
`other-extensions` and adds explicit `{-# LANGUAGE #-}` pragmas
to the respective modules.

This also removes a few obsolete pre-Haddock-2 `#hide` pragmas.

Note: The new `default-extensions` field lists extensions that are used
in all modules in the component, while the `other-extensions` field
lists extensions that are used in some modules, e.g. via the
`{-# LANGUAGE #-}` pragma.

Signed-off-by: Herbert Valerio Riedel <hvr@gnu.org>