Revert "[Cmm Sized] Adjust"
In old GHC:

    I8# :: Int# -> Int8
    indexInt8OffAddr# :: Addr# -> Int# -> Int#
    ...

In upcoming GHC 9.2:

    I8# :: Int8# -> Int8
    indexInt8OffAddr# :: Addr# -> Int# -> Int8#
    ...

Both of those work with the last array release, where the only
requirement is merely that the constructor and primops agree, i.e.:

    exists alpha.
    I8# :: alpha -> Int8
    indexInt8OffAddr# :: Addr# -> Int# -> alpha

In current GHC HEAD, we had

    I8# :: Int8# -> Int8
    indexInt8OffAddr# :: Addr# -> Int# -> Int#

and it was only because `Int8# /= Int#` that we needed the commit being
reverted. we are about to fix the primops to match the constructors in
accordance with the final 9.2 design, and so we don't need that commit
anymore.

This reverts commit c7a696e3e6d5a6b00d3e00ca694af916f15bcff5.